Rename TM deliveryservice to ds
