+++ traffic_monitor_golang/traffic_monitor/ds/stat.go
@@ -1,4 +1,4 @@
-package deliveryservice
+package ds
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
